Made some CSS changes, and created most of the play game file.

+++ backend/backendapi/databaseapi.php
@@ -1,5 +1,6 @@
 <?php
 include dirname(__DIR__, 1)."/objects/database.php";
+//include dirname(__DIR__, 1).'/objects/error.php';
 include dirname(__DIR__, 1).'/objects/result.php';
 
 
@@ -11,15 +12,13 @@ class DatabaseAPI{
     $this->conn = $database->getConnection();
   }
 
-  // ******** USER FUNCTIONS ********
-
-  /* getUserPassword(string):
-  parameter: valid username
-  returns:
-   - Error if the query fails
-   - String if supplied a valid username
-   - False bool if username does not exist
-  */
+/* getUserPassword(string):
+parameter: valid username
+returns:
+ - Error if the query fails
+ - String if supplied a valid username
+ - False bool if username does not exist
+*/
   public function getUserPassword($uName){
     $result = new Result();
     $findUser = "SELECT pWord FROM user WHERE uName = '$uName'";
@@ -47,6 +46,7 @@ class DatabaseAPI{
       }
     }
   }
+
   /* checkUsername(string):
   parameter: valid username
   returns:
@@ -77,6 +77,7 @@ class DatabaseAPI{
       }
     }
   }
+
   /* removeUser(string):
   parameter: valid username
   returns:
@@ -109,6 +110,7 @@ class DatabaseAPI{
       return $doesUserExist;
     }
   }
+
   /* createUser(string, string):
   parameter: valid username, valid password
   returns:
@@ -145,43 +147,12 @@ class DatabaseAPI{
       return $doesUserExist;
     }
   }
-  // QUESTION:: do we need to find out all of the things linked to this user and delete them as well?
 
-  public function getUserID_Username($uName){
-    $result = new Result();
-    $findUser = "SELECT uID FROM user WHERE uName = '$uName'";
-    $query = mysqli_query($this->conn, $findUser);
 
-    if (!$query){ // If something went wrong with the query, return appropriate error
-      // Error return: Unable to query DB
-      $result->setError("DatabaseAPI Error: mysqli_query error");
-      return $result;
-    }
-    else{
-      $rows = mysqli_num_rows($query); // Find number of rows retrieved
-      // If the user exists, only ONE row should be returned, else user does not exist.
-      if($rows == 1){
-        $uID = mysqli_fetch_assoc($query);
-        $result->setResult($uID["uID"]);
-        return $result;
-      }
-      else{
-        $result->setError("DatabaseAPI Error: Duplicate UName in DB");
-        return $result;
-      }
-    }
-  }
-  // ******** PROJECT FUNCTIONS ********
-
-  /* createProject(string, int)
-  parameters: valid project name, valid ownerID
-  returns:
-   - error if query fails
-   - true if creation succeeds
-  */
   public function createProject($projectName, $ownerID){
     $result = new Result();
-    $addProject = "INSERT INTO `project`(`projectName`, `ownerID`, `isPublished`) VALUES ('$projectName','$ownerID',0)";
+    $addProject = "INSERT INTO project(projectName, ownerID) VALUES ('$projectName', '$ownerID')";
+
     $query = mysqli_query($this->conn, $addProject);
 
     if (!$query){
@@ -193,34 +164,29 @@ class DatabaseAPI{
       return $result;
     }
   }
-  /* removeProject(int)
-  parameters: valid ownerID
-  returns:
-   - error if query fails
-   - true if deletion succeeds
-  */
+
   public function removeProject($projectID){
     $result = new Result();
+    if ($this->checkUsername($uName)){
+      $deleteProject = "DELETE FROM project WHERE projectID = " . $projectID;
+      $query = mysqli_query($this->conn, $deleteProject);
 
-    $deleteProject = "DELETE FROM project WHERE projectID = " . $projectID;
-    $query = mysqli_query($this->conn, $deleteProject);
-
-    if (!$query){ // If something went wrong with the query, return appropriate error
-      // Error return: Unable to query DB
-      $result->setError("Error: mysqli_query error");
-      return $result;
+      if (!$query){ // If something went wrong with the query, return appropriate error
+        // Error return: Unable to query DB
+        $result->setError("Error: mysqli_query error");
+        return $result;
+      }
+      else{
+        $result->setResult(true);
+        return $result;
+      }
     }
-    else{
-      $result->setResult(true);
+    else {
+      $result->setResult(false);
       return $result;
     }
   }
-  /* getAllPublishedProjects()
-  parameters: n/a
-  returns:
-   - error if query fails
-   - set of all published projects
-  */
+
   public function getAllPublishedProjects(){
     $result = new Result();
     $findUser = "SELECT * FROM project WHERE isPublished = 1";
@@ -236,106 +202,26 @@ class DatabaseAPI{
       return $result;
     }
   }
-
-  public function getAllProjects($username){
-
-    $uID = $this->getUserID_Username($username);
-
-    if ($uID->isError()){
-      return $uID;
-    }
-    else{
+    
+  public function getImg(){
       $result = new Result();
-      $findUser = "SELECT * FROM project WHERE ownerID = ".$uID->getResult()."";
-      $query = mysqli_query($this->conn, $findUser);
-
-      if (!$query){ // If something went wrong with the query, return appropriate error
-        // Error return: Unable to query DB
-        $result->setError("Error: mysqli_query error");
-        return $result;
-      }
-      else{
-        $result->setResult($query);
-        return $result;
-      }
-    }
-
-
-  }
-  // ******** SCENE FUNCTIONS ********
-
-  /* checkSceneID(int):
-  parameter: valid sceneID
-  returns:
-   - Error if the query fails
-   - true bool if supplied an existing sceneID
-   - false bool if supplied a sceneID that does not exist
-  */
-  public function checkSceneID($sceneID){
-    $result = new Result();
-    $findUser = "SELECT sceneID FROM scene WHERE sceneID = '$sceneID'";
-    $query = mysqli_query($this->conn, $findUser);
-
-    if (!$query){ // If something went wrong with the query, return appropriate error
-      // Error return: Unable to query DB
-      $result->setError("DatabaseAPI Error: mysqli_query error");
+      $image = mysqli_query($this->conn, "SELECT imgDir FROM scene WHERE projectID=1 AND sceneID=1");
+      $result->setResult($image);
       return $result;
-    }
-    else{
-      $rows = mysqli_num_rows($query); // Find number of rows retrieved
-      // If the user exists, only ONE row should be returned, else user does not exist.
-      if($rows == 1){
-        $result->setResult(true);
-        return $result;
-      }
-      else{
-        $result->setResult(false);
-        return $result;
-      }
-    }
   }
-  /* removeScene(int)
-  parameter: valid sceneID
-  return:
-   - error if query fails
-   - true bool if successful
-  */
-  public function removeScene($sceneID){
-    $result = new Result();
-    $doesSceneExist = $this->checkSceneID($sceneID);
-    if ((!$doesSceneExist->getError()) && (($doesSceneExist->getResult())==true)){
-      $removeScene = "DELETE FROM sceneID WHERE sceneID = '$sceneID'";
-      $query = mysqli_query($this->conn, $removeScene);
-
-      if (!$query){
-        $result->setError("DatabaseAPI Error: mysqli_query error");
-        return $result;
-      }
-      else{
-        $SQL = "DELETE FROM `options` WHERE sceneID LIKE $sceneID";
-        $query = mysqli_query($this->conn, $removeScene);
-        if (!$query){
-          $result->setError("DatabaseAPI Error: mysqli_query error");
-          return $result;
-        }
-        else {
-          $result->setResult(true);
-          return $result;
-        }
-      }
-    }
-    else if ((!$doesSceneExist->getError()) && (($doesSceneExist->getResult())==false)){
-      $result->setError("DatabaseAPI Error: sceneID does not exist, and cannot be deleted from DB");
+    
+  public function getTxt(){
+      $result = new Result();
+      $txt = mysqli_query($this->conn, "SELECT description FROM scene WHERE projectID=1 AND sceneID=1");
+      $result->setResult($txt);
       return $result;
-    }
-    else{
-      return $doesSceneExist;
-    }
   }
-
-
-
-
+    
+  public function getOptions(){
+      $result = new Result();
+      $opts = mysqli_query($this->conn, "SELECT oText FROM options WHERE sceneID=1");
+      $result->setResult($opts);
+      return $result;
+  }
 }
-
 ?>
